pyproject.toml
[project]
name = "mcp-server-openai"
version = "0.1.0"
description = "MCP server for OpenAI API integration"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.1",
    "openai>=1.0.0",
    "click>=8.0.0",
    "pytest-asyncio"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-openai = "mcp_server_openai.server:main"